requirements.txt
Flask==2.3.2
PyMySQL==1.1.1
Werkzeug==2.3.6
python-dotenv==1.0.0
Flask-Session==0.5.0
